# LibraryManagementSystem_API/Business/Dtos/CommentDto/PostCommentDto.cs
namespace LibraryManagementSystem_API.Business.Dtos.CommentDto
{
    public class PostCommentDto
    {
 /*       public Guid BookId { get; set; }*/
        public string? Username { get; set; }
        public float Star { get; set; }
        public string? Comment { get; set; }

/*        public DateTime? CreatedDate { get; set;}*/
    }
}