# LibraryManagementSystem_API/Business/Dtos/RequestDtos/PostRequestDto.cs
namespace LibraryManagementSystem_API.Business.Dtos.RequestDtos
{
    public class PostRequestDto
    {
        public string Username {  get; set; }

        public string Type { get; set; }

    }
}